requirements.txt
numpy==1.24.4
Pillow==10.1.0
tenseal==0.3.14